Input Singleton Enhancements (#83)
* Add `mouse_click`, `mouse_unlick`, `mouse_pos`, `mouse_offset`, `mouse_scrolled`, `window_resized`, `window_width` and `window_height` to the `InputSingleton`.
* Removed the `Window*` from all systems, now that data is accessed through the singleton.
* Removed `Scene::Clear`, just recreate the scene from scratch when needed.
* Add `Scene::post_update` for resetting parts of the input singleton that contain per-frame data.
* The `Loader` no longer adds the singleton, that is done in the scene constructor.
* `CameraSystem` is now stateless.

+++ Sprocket/Utility/Maths.cpp
@@ -67,7 +67,7 @@ glm::mat4 NoScale(const glm::mat4& matrix)
     return Maths::Transform(pos, ori);
 }
 
-glm::vec3 GetMouseRay(const glm::vec2& mousePos, u32 w, u32 h, const glm::mat4& view, const glm::mat4& proj)
+glm::vec3 GetMouseRay(const glm::vec2& mousePos, float w, float h, const glm::mat4& view, const glm::mat4& proj)
 {
     // Homogeneous Clip Space
     glm::vec4 ray = {(2.0f * mousePos.x) / w - 1, -((2.0f * mousePos.y) / h - 1), -1.0f, 1.0f};
